// src/Bottleneko.Client/src/lib/api/bottleneko.gen.ts
export type ScriptStatus
    = 'Stopped'
        | 'Starting'
        | 'Running'
        | 'Restarting'
        | 'Stopping'
        | 'Error';
export const ScriptStatusValues : ScriptStatus[] = [
    'Stopped',
    'Starting',
    'Running',
    'Restarting',
    'Stopping',
    'Error',
];

export type ScriptCode
    = JsScriptCode;

export interface JsScriptCode {
    $type: 'JavaScript';
    source: string;
}

export interface ScriptDto {
    id: string;
    name: string;
    description: string;
    autoStart: boolean;
    code: ScriptCode;
    status: ScriptStatus;
}